renaming various folders, files, and functions
