+++ qiskit_dynamics/pulse/backend_parser/operator_from_string.py
